Create class SoundSplit
Make the logic of auto split easier to UnitTest and reusable
Add comment to the class to explain what it does
